+ added support for post-login skill training notification ~ fixed more issues with the skill queue not being synced up or improperly notifying the client ~ moved skill queue code out of the Character class into the skillMgr bound service ~ simplified skillMgr code as much as possible

+++ Server/Node/Inventory/Items/Types/Station.cs
@@ -48,7 +48,7 @@ namespace Node.Inventory.Items.Types
         public double DockingCostPerVolume => mDockingCostPerVolume;
         public double MaxShipVolumeDockable => mMaxShipVolumeDockable;
         public int OfficeRentalCost => mOfficeRentalCost;
-        public int SolarSystemID => (int) this.LocationID;
+        public int SolarSystemID => this.LocationID;
         public int ConstellationID => mConstellationID;
         public int RegionID => mRegionID;
         public double ReprocessingEfficiency => mReprocessingEfficiency;
